ectracted report generating and report processing out of index.js
