+++ scanners/web-endpoint/index.js
@@ -12,6 +12,7 @@ const {
   GRAPHQL_URL,
   NATS_SUB_STREAM,
 } = process.env;
+
 // NATs connection 
 const nc = await connect({ servers: NATS_URL, })
 const jc = JSONCodec()
@@ -35,40 +36,8 @@ process.on('SIGINT', () => process.exit(0))
       if (await isWebEndpointType(endpoint, pageInstance)) { //filtering for only web endpoints
         // Get pages for webendpoints (with url slugs)
         const pages = await getPages(endpoint, pageInstance, browser);
+        const accessibilityPages = await processAxeReport(pageInstance, pages, browser)
 
-// -------------------------------------------------------------------------------------
-        let webEndpointAxeResults = {}  //form response
-
-        for (const pageToEvaluate of pages) {
-          console.log('Evaluating page: ', pageToEvaluate)
-          const axeReport = await evaluateAccessibility(pageToEvaluate, pageInstance, browser)
-
-          // Process report (create camelCase key for each evaluated criterion )
-          for (let i = 0; i < axeReport.length; i++) {
-            const camelize = s => s.replace(/-./g, x => x[1].toUpperCase())
-            const criterion = axeReport[i];
-            const criterionKey = Object.keys(criterion )[0];
-            const criterionValue = criterion [criterionKey];
-            const criterionKeyCamelCase = camelize(criterionKey);
-            if (!webEndpointAxeResults[pageToEvaluate]) {
-              webEndpointAxeResults[pageToEvaluate] = {}
-            }
-            if (typeof criterionValue.checkPasses === 'boolean') {
-              criterionValue.checkPasses = criterionValue.checkPasses.toString()
-            }
-            webEndpointAxeResults[pageToEvaluate][criterionKeyCamelCase] = {
-              checkPasses: criterionValue.checkPasses,
-              metadata: criterionValue.metadata,
-            }
-          }
-
-        }
-        const accessibilityPages = Object.keys(webEndpointAxeResults).map(page => {
-          return {
-            url: page,
-            ...webEndpointAxeResults[page],
-          }
-        })
         const mutation = gql`
             mutation {
               webEndpoint(
@@ -80,14 +49,19 @@ process.on('SIGINT', () => process.exit(0))
               )
             }
             `;
+        // console.log('*****************************************************\n', mutation)
+
         // API connection 
         const graphQLClient = new GraphQLClient(GRAPHQL_URL);
+        
         // Write mutation to GraphQL API
-        // console.log('*****************************************************\n', mutation)
         const mutationResponse = await graphQLClient.request(mutation);
+        
         console.info("wrote mutation to GraphQL API with response", mutationResponse);
+        
         await pageInstance.close()
-      }
+      } // end of if web endpoint
+
       console.log(`Accessibility scan of ${endpoint} complete.`)
       console.log('-----')
     }
